remove all non translate functionality
