+++ packages/i18n/package.json
@@ -34,8 +34,6 @@
   "module": "index.mjs",
   "esnext": "index.esnext",
   "dependencies": {
-    "@shopify/dates": "^2.0.1",
-    "@shopify/decorators": "^3.0.1",
     "@shopify/function-enhancers": "^3.0.1"
   },
   "exports": {
