--- packages/i18n/src/currencyCode.ts
@@ -1,161 +1,0 @@
-export enum CurrencyCode {
-  Usd = 'USD',
-  Eur = 'EUR',
-  Gbp = 'GBP',
-  Cad = 'CAD',
-  Afn = 'AFN',
-  All = 'ALL',
-  Dzd = 'DZD',
-  Aoa = 'AOA',
-  Ars = 'ARS',
-  Amd = 'AMD',
-  Awg = 'AWG',
-  Aud = 'AUD',
-  Bbd = 'BBD',
-  Azn = 'AZN',
-  Bdt = 'BDT',
-  Bsd = 'BSD',
-  Bhd = 'BHD',
-  Bif = 'BIF',
-  Byn = 'BYN',
-  Bzd = 'BZD',
-  Bmd = 'BMD',
-  Btn = 'BTN',
-  Bam = 'BAM',
-  Brl = 'BRL',
-  Bob = 'BOB',
-  Bwp = 'BWP',
-  Bnd = 'BND',
-  Bgn = 'BGN',
-  Mmk = 'MMK',
-  Khr = 'KHR',
-  Cve = 'CVE',
-  Kyd = 'KYD',
-  Xaf = 'XAF',
-  Clp = 'CLP',
-  Cny = 'CNY',
-  Cop = 'COP',
-  Kmf = 'KMF',
-  Cdf = 'CDF',
-  Crc = 'CRC',
-  Hrk = 'HRK',
-  Czk = 'CZK',
-  Dkk = 'DKK',
-  Djf = 'DJF',
-  Dop = 'DOP',
-  Xcd = 'XCD',
-  Egp = 'EGP',
-  Ern = 'ERN',
-  Etb = 'ETB',
-  Fkp = 'FKP',
-  Xpf = 'XPF',
-  Fjd = 'FJD',
-  Gmd = 'GMD',
-  Ghs = 'GHS',
-  Gip = 'GIP',
-  Gtq = 'GTQ',
-  Gnf = 'GNF',
-  Gyd = 'GYD',
-  Gel = 'GEL',
-  Htg = 'HTG',
-  Hnl = 'HNL',
-  Hkd = 'HKD',
-  Huf = 'HUF',
-  Isk = 'ISK',
-  Inr = 'INR',
-  Idr = 'IDR',
-  Irr = 'IRR',
-  Ils = 'ILS',
-  Iqd = 'IQD',
-  Jmd = 'JMD',
-  Jpy = 'JPY',
-  Jep = 'JEP',
-  Jod = 'JOD',
-  Kid = 'KID',
-  Kzt = 'KZT',
-  Kes = 'KES',
-  Kwd = 'KWD',
-  Kgs = 'KGS',
-  Lak = 'LAK',
-  Lvl = 'LVL',
-  Lbp = 'LBP',
-  Lsl = 'LSL',
-  Lrd = 'LRD',
-  Lyd = 'LYD',
-  Ltl = 'LTL',
-  Mga = 'MGA',
-  Mkd = 'MKD',
-  Mop = 'MOP',
-  Mwk = 'MWK',
-  Mvr = 'MVR',
-  Mxn = 'MXN',
-  Myr = 'MYR',
-  Mru = 'MRU',
-  Mur = 'MUR',
-  Mdl = 'MDL',
-  Mad = 'MAD',
-  Mnt = 'MNT',
-  Mzn = 'MZN',
-  Nad = 'NAD',
-  Npr = 'NPR',
-  Ang = 'ANG',
-  Nzd = 'NZD',
-  Nio = 'NIO',
-  Ngn = 'NGN',
-  Nok = 'NOK',
-  Omr = 'OMR',
-  Pkr = 'PKR',
-  Pab = 'PAB',
-  Pgk = 'PGK',
-  Pyg = 'PYG',
-  Pen = 'PEN',
-  Php = 'PHP',
-  Pln = 'PLN',
-  Qar = 'QAR',
-  Ron = 'RON',
-  Rub = 'RUB',
-  Rwf = 'RWF',
-  Wst = 'WST',
-  Shp = 'SHP',
-  Sar = 'SAR',
-  Std = 'STD',
-  Rsd = 'RSD',
-  Scr = 'SCR',
-  Sll = 'SLL',
-  Sgd = 'SGD',
-  Sos = 'SOS',
-  Sdg = 'SDG',
-  Syp = 'SYP',
-  Zar = 'ZAR',
-  Krw = 'KRW',
-  Ssp = 'SSP',
-  Sbd = 'SBD',
-  Lkr = 'LKR',
-  Srd = 'SRD',
-  Szl = 'SZL',
-  Sek = 'SEK',
-  Chf = 'CHF',
-  Tjs = 'TJS',
-  Twd = 'TWD',
-  Thb = 'THB',
-  Tzs = 'TZS',
-  Top = 'TOP',
-  Ttd = 'TTD',
-  Tnd = 'TND',
-  Try = 'TRY',
-  Tmt = 'TMT',
-  Ugx = 'UGX',
-  Uah = 'UAH',
-  Aed = 'AED',
-  Uyu = 'UYU',
-  Uzs = 'UZS',
-  Vuv = 'VUV',
-  Ves = 'VES',
-  Vnd = 'VND',
-  Xof = 'XOF',
-  Yer = 'YER',
-  Zmw = 'ZMW',
-  // Codes after this line are deprecated
-  Byr = 'BYR',
-  Vef = 'VEF',
-}
