+++ packages/i18n/src/index.ts
@@ -1,2 +1,7 @@
 export * from './locale';
 export * from './pseudotranslate';
+export * from './simpleI18n';
+export type {
+  TranslationDictionary,
+  ReplacementDictionary,
+} from './simpleI18nUtils/types';
