+++ packages/i18n/src/simpleI18nUtils/index.ts
@@ -1,0 +1,1 @@
+export {translate, translationKeyExists} from './translate';
