+++ packages/i18n/src/simpleI18nUtils/translate.ts
@@ -1,9 +1,7 @@
 import {memoize as memoizeFn} from '@shopify/function-enhancers';
 
-import {TranslationDictionary, PrimitiveReplacementDictionary} from '../types';
-import {MissingTranslationError, MissingReplacementError} from '../errors';
-
-import {DEFAULT_FORMAT} from './interpolate';
+import {TranslationDictionary, ReplacementDictionary} from './types';
+import {MissingTranslationError, MissingReplacementError} from './errors';
 
 const MISSING_TRANSLATION = Symbol('Missing translation');
 const CARDINAL_PLURALIZATION_KEY_NAME = 'count';
@@ -15,6 +13,13 @@ const LATIN = 'latn';
 const isString = (value: any): value is string => typeof value === 'string';
 
 const numberFormats = new Map<string, Intl.NumberFormat>();
+
+/**
+ * Used to interpolate a string with values from an object.
+ * `{key}` will be replaced with `replacements[key]`.
+ */
+export const INTERPOLATE_FORMAT = /{\s*(\w+)\s*}/g;
+
 export function memoizedNumberFormatter(
   locales?: string | string[],
   options?: Intl.NumberFormatOptions,
@@ -75,9 +80,35 @@ export const memoizedPluralRules = memoizeFn(
     `${locale}${JSON.stringify(options)}`,
 );
 
+export function translationKeyExists(
+  id: string,
+  translations: TranslationDictionary | TranslationDictionary[],
+): boolean {
+  const normalizedTranslations = Array.isArray(translations)
+    ? translations
+    : [translations];
+
+  let result: string | TranslationDictionary;
+
+  for (const translationDictionary of normalizedTranslations) {
+    result = translationDictionary;
+
+    for (const part of id.split(SEPARATOR)) {
+      result = result[part];
+      if (!result) break;
+    }
+
+    if (result) {
+      return true;
+    }
+  }
+
+  return false;
+}
+
 export function translate(
   id: string,
-  replacements: PrimitiveReplacementDictionary,
+  replacements: ReplacementDictionary,
   translations: TranslationDictionary | TranslationDictionary[],
   locale: string,
 ): string {
@@ -105,7 +136,7 @@ function translateWithDictionary(
   id: string,
   translations: TranslationDictionary,
   locale: string,
-  replacements?: PrimitiveReplacementDictionary,
+  replacements?: ReplacementDictionary,
 ): string | typeof MISSING_TRANSLATION {
   let result: string | TranslationDictionary = translations;
 
@@ -177,11 +208,11 @@ function translateWithDictionary(
 
 function updateStringWithReplacements(
   str: string,
-  replacements: PrimitiveReplacementDictionary = {},
+  replacements: ReplacementDictionary = {},
 ): string {
   const pieces: string[] = [];
 
-  const replaceFinder = new RegExp(DEFAULT_FORMAT, 'g');
+  const replaceFinder = new RegExp(INTERPOLATE_FORMAT, 'g');
 
   let lastOffset = 0;
