+++ packages/i18n/src/simpleI18nUtils/types.ts
@@ -1,0 +1,7 @@
+export interface TranslationDictionary {
+  [key: string]: string | TranslationDictionary;
+}
+
+export interface ReplacementDictionary {
+  [key: string]: string | number;
+}
