--- packages/i18n/src/utilities/interpolate.ts
@@ -1,5 +1,0 @@
-/**
- * Used to interpolate a string with values from an object.
- * `{key}` will be replaced with `replacements[key]`.
- */
-export const DEFAULT_FORMAT = /{\s*(\w+)\s*}/g;
