--- packages/i18n/src/utilities/money.ts
@@ -1,38 +1,0 @@
-import {DIRECTION_CONTROL_CHARACTERS} from '../constants';
-
-import {memoizedNumberFormatter} from './translate';
-
-export function getCurrencySymbol(
-  locale: string,
-  options: Intl.NumberFormatOptions,
-) {
-  const currencyStringRaw = formatCurrency(0, locale, options);
-  const controlChars = new RegExp(`[${DIRECTION_CONTROL_CHARACTERS}]*`, 'gu');
-  const currencyString = currencyStringRaw.replace(controlChars, '');
-  const matchResult = /\p{Nd}\p{Po}*\p{Nd}*/gu.exec(currencyString);
-  if (!matchResult) {
-    throw new Error(
-      `Number input in locale ${locale} is currently not supported.`,
-    );
-  }
-  const formattedAmount = matchResult[0];
-  const [currencyPrefix, currencySuffix] =
-    currencyString.split(formattedAmount);
-  const elements = {
-    symbol: currencyPrefix || currencySuffix,
-    prefixed: currencyPrefix !== '',
-  };
-
-  return elements;
-}
-
-export function formatCurrency(
-  amount: number,
-  locale: string,
-  options: Intl.NumberFormatOptions,
-) {
-  return memoizedNumberFormatter(locale, {
-    style: 'currency',
-    ...options,
-  }).format(amount);
-}
