--- packages/i18n/src/utilities/string.ts
@@ -1,3 +1,0 @@
-export function convertFirstSpaceToNonBreakingSpace(str: string) {
-  return str.replace(' ', '\u00A0');
-}
