+++ packages/i18n/tsconfig.json
@@ -9,5 +9,8 @@
     "../../config/typescript/*.d.ts",
     "./src/**/*.ts",
     "./src/**/*.tsx"
+  ],
+  "references": [
+    {"path": "../function-enhancers"}
   ]
 }
